Move options to consts
