+++ scripts/utils.ts
@@ -1,6 +1,5 @@
 import { access, readFile, writeFile } from 'node:fs/promises';
 import { parse } from 'jsonc-parser';
-import { type PackageManagerName, packageManagers } from 'nypm';
 
 export const exists = async (path: string) => {
   try {
@@ -41,21 +40,6 @@ export const title = `
 Y88b. .d88P 888        888     888  T88b   d8888888888 Y88b  d88P  888       888     888        
  "Y88888P"  88888888   888     888   T88b d88P     888  "Y8888P" 8888888     888     8888888888
 `;
-
-export const options = {
-  packageManagers: packageManagers.map((pm) => pm.name) as PackageManagerName[],
-  editorConfigs: ['vscode', 'zed'],
-  editorRules: [
-    'vscode-copilot',
-    'cursor',
-    'windsurf',
-    'zed',
-    'claude',
-    'codex',
-    'kiro',
-  ],
-  integrations: ['husky', 'lefthook', 'lint-staged'],
-};
 
 export const updatePackageJson = async ({
   dependencies,
